update controller and service

+++ src/main/java/com/example/demo/application/service/ProductService.java
@@ -4,7 +4,9 @@ import com.example.demo.domain.model.Product;
 import com.example.demo.domain.repository.ProductRepository;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -21,12 +23,16 @@ public class ProductService {
         return productRepository.findAll();
     }
 
-    public Product getProductById(Long id) {
+    public Product getById(Long id) {
+        // Logika Optional dikelola di sini
         return productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Produk tidak ditemukan dengan id: " + id));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Produk tidak ditemukan"));
     }
 
-    public void deleteProduct(Long id) {
+    public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Gagal menghapus! Produk tidak ditemukan");
+        }
         productRepository.deleteById(id);
     }
 }
